==> Code/3_Linear Model.ipynb <==
import numpy as np
import pandas as pd
import pickle as pkl
import matplotlib.pyplot as plt
import nltk
from nltk.tokenize import sent_tokenize, word_tokenize
import string

games = pd.read_csv('switch_game_info_filtered.csv')
critic_reviews = pd.read_csv('switch_critic_review.csv')
user_reviews = pd.read_csv('switch_user_review.csv')
sales = pd.read_csv('sales_switch.csv')

Merge databases

games_sales = pd.merge(left=games, right=sales, how='inner', on='title')

Apply data engineering for developers and filter required columns

fir_party = ['Nintendo','Nintendo EAD Tokyo , Nintendo Software Technology',
             'Retro Studios','Intelligent Systems, Koei Tecmo Games',
             'HAL Labs','Nintendo, Nd Cube', 'Monolith Soft']
sec_party = ['PlatinumGames','Nintendo, PlatinumGames','Next Level Games, Nintendo','Asobo Studio, Engine Software',
             'Camelot Software Planning','Nintendo, GREZZO','Nintendo, HAL Labs, Bandai Namco Games, Sora Ltd.',
             'indieszero']

games_sales['Developer_type'] = 'Third'
games_sales.loc[games_sales['developer'].isin(fir_party), 'Developer_type'] = 'First'
games_sales.loc[games_sales['developer'].isin(sec_party) , 'Developer_type'] = 'Second'



games_sales = games_sales[['total_sales_USDMM',
                           'rating', 'meta_overview',
                           'user_overview', 'genre_action', 'genre_adventure', 'genre_fighting',
                           'genre_platform', 'genre_puzzle', 'genre_racing', 'genre_roleplay',
                           'genre_shooter', 'genre_simulation', 'genre_sports', 'genre_strategy',
                           'genre_misc', 'Developer_type']]
games_sales

First model with all features

import statsmodels.api as sm
from scipy import stats

games_sales_dum = pd.get_dummies(games_sales, columns = ['rating','meta_overview','user_overview','Developer_type'])

# Setup dependent and independent variables
X=games_sales_dum.iloc[:,1:]
y=games_sales_dum['total_sales_USDMM']


X2 = sm.add_constant(X)
est = sm.OLS(y, X2)
est2 = est.fit()
print(est2.summary())

Final model with removed features

X_filt = X.copy()
X_filt = X_filt.drop(['genre_roleplay','rating_T','Developer_type_Second',
                      'genre_fighting','genre_adventure','genre_platform', 'genre_puzzle',
                      'genre_sports','genre_misc','genre_strategy','genre_action',
                      'rating_E10+',
                      'Developer_type_Third'],axis=1)

X2 = sm.add_constant(X_filt)
est = sm.OLS(y, X2)
est2 = est.fit()
print(est2.summary())
